+++ resources/views/livewire/resource/create.blade.php
@@ -1,18 +1,18 @@
 <div class="">
-    @section('title', __('Create ' . $model->singularName()))
+    @section('title', $model->createHeaderTitle())
 
 
     @if(!$inModal)
     <x-aura::breadcrumbs>
         <x-aura::breadcrumbs.li :href="route('aura.dashboard')" title="" icon="dashboard" iconClass="text-gray-500 w-6 h-6 mr-0" />
-        <x-aura::breadcrumbs.li :href="route('aura.' . $model->getSlug() . '.index')" :title="__($model->getPluralName())" />
-        <x-aura::breadcrumbs.li title="{{ __('Create :resource', ['resource' => __($model->singularName())]) }}" />
+        <x-aura::breadcrumbs.li :href="route($model->createReturnRoute())" :title="__($model->getPluralName())" />
+        <x-aura::breadcrumbs.li :title="$model->createHeaderTitle()" />
     </x-aura::breadcrumbs>
     @endif
 
     <div class="flex items-center justify-between {{ $inModal ? 'mb-8' : 'my-8'}}">
         <div>
-            <h1 class="text-2xl font-semibold">{{ __('Create :resource', ['resource' => __($model->singularName())]) }}</h1>
+            <h1 class="text-2xl font-semibold">{{ $model->createHeaderTitle() }}</h1>
         </div>
 
         @if(!$inModal)